Extract getHistory function into helper
