+++ src/components/LineChart.js
@@ -1,9 +1,10 @@
 import PropTypes from 'prop-types';
 import { Line } from 'react-chartjs-2';
-import axios from 'axios';
 import { useState, useEffect } from 'react';
+import getHistory from '../helpers/getHistory';
 
 const LineChart = (props) => {
+  const { coinName, fiatName, coinId } = props;
   const initialState = {
     labels: ['a', 'b'],
     datasets: [
@@ -11,27 +12,6 @@ const LineChart = (props) => {
     ],
   };
   const [data, setData] = useState(initialState);
-
-  const getHistory = async (coinId) => {
-    const res = await axios.get(`https://api.coinstats.app/public/v1/charts?period=1m&coinId=${coinId}`);
-    const usdPrices = res.data.chart.map((array) => array[1]);
-    const UNIXDates = res.data.chart.map((array) => array[0]);
-    const dates = UNIXDates.map((date) => {
-      const newDate = new Date(date * 1000);
-      return newDate.toLocaleString('en-US', { month: 'short', day: 'numeric' });
-    });
-    const dataRes = {
-      labels: dates,
-      datasets: [
-        {
-          label: 'CURRENCY PRICE IN USD',
-          data: usdPrices,
-          borderColor: '#ff922f',
-        },
-      ],
-    };
-    setData(dataRes);
-  };
 
   const options = {
     responsive: true,
@@ -53,10 +33,9 @@ const LineChart = (props) => {
     },
   };
 
-  const { coinId } = props;
   useEffect(() => {
-    getHistory(coinId);
-  }, [coinId]);
+    getHistory(coinId, setData, coinName, fiatName);
+  }, [fiatName]);
 
   return (
     <div style={{
@@ -73,10 +52,14 @@ const LineChart = (props) => {
 
 LineChart.propTypes = {
   coinId: PropTypes.string,
+  coinName: PropTypes.string,
+  fiatName: PropTypes.string,
 };
 
 LineChart.defaultProps = {
   coinId: 'bitcoin',
+  coinName: 'Bitcoin',
+  fiatName: 'USD',
 };
 
 export default LineChart;
