+++ src/helpers/getHistory.js
@@ -1,0 +1,24 @@
+import axios from 'axios';
+
+const getHistory = async (coinId, callback, coin, fiat) => {
+  const res = await axios.get(`https://api.coinstats.app/public/v1/charts?period=1m&coinId=${coinId}`);
+  const usdPrices = res.data.chart.map((array) => array[1]);
+  const UNIXDates = res.data.chart.map((array) => array[0]);
+  const dates = UNIXDates.map((date) => {
+    const newDate = new Date(date * 1000);
+    return newDate.toLocaleString('en-US', { month: 'short', day: 'numeric' });
+  });
+  const dataRes = {
+    labels: dates,
+    datasets: [
+      {
+        label: `${coin.toUpperCase()} PRICE IN ${fiat.toUpperCase()}`,
+        data: usdPrices,
+        borderColor: '#ff922f',
+      },
+    ],
+  };
+  callback(dataRes);
+};
+
+export default getHistory;
